feat: refactor quantity to use BigDecimal for precision in Asset and Operation classes - quantity can be a float number (split, inplit, etfs etc)

+++ src/main/java/com/estevam/calculacoes/operation/Operation.java
@@ -12,10 +12,10 @@ public class Operation {
     private String assetCode;
     private OperationType type;
     private String marketType;
-    private int quantity;
+    private BigDecimal quantity;
     private BigDecimal price;
 
-    public Operation(LocalDate date, String assetCode, OperationType type, String marketType, int quantity, BigDecimal price) {
+    public Operation(LocalDate date, String assetCode, OperationType type, String marketType, BigDecimal quantity, BigDecimal price) {
         this.date = date;
         this.assetCode = assetCode.toUpperCase();
         this.type = type;
@@ -34,5 +34,4 @@ public class Operation {
                 ", price=" + price +
                 '}';
     }
-
 }
